Problemas resolvidos

+++ src/main/java/Mapped/api/models/entities/Login.java
@@ -2,20 +2,17 @@ package Mapped.api.models.entities;
 
 public class Login {
 
-
-    private  int id;
-    private int CPF;
+    private int id;
+    private long CPF;
     private String Senha;
-
 
     public Login() {
     }
 
-    public Login(int id, int CPF, String senha) {
+    public Login(int id, long CPF, String senha) {
         this.id = id;
         this.CPF = CPF;
         this.Senha = senha;
-
     }
 
     public int getId() {
@@ -26,11 +23,11 @@ public class Login {
         this.id = id;
     }
 
-    public int getCPF() {
+    public long getCPF() {
         return CPF;
     }
 
-    public void setCPF(int CPF) {
+    public void setCPF(long CPF) {
         this.CPF = CPF;
     }
